fix(lexer): support decimal point in numeric literals
Numbers like 123.45 were tokenized as "123" + unknown "." + "45".
Lexer.number() now consumes a decimal point followed by digits.

Since TM1 numbers are all IEEE-754 doubles, the parser now always
parses NUMBER tokens as float instead of int/float depending on the
decimal point.

Added lexer and parser tests for decimal numbers.

Assisted by Claude Code

+++ src/linti/parser/parser.py
@@ -558,8 +558,8 @@ class Parser:
         # Number literal
         if tok.type == TokenType.NUMBER:
             self.advance()
-            # keep as int for now; extend later for float
-            return Number(int(tok.value), token=tok)
+            # TM1 has no integer type; all numbers are IEEE-754 doubles.
+            return Number(float(tok.value), token=tok)
 
         # String literal
         if tok.type == TokenType.STRING:
